Correct abs function should be used (clang -Wabsolute-value)

+++ src/iop/basecurve.c
@@ -415,8 +415,8 @@ dt_iop_basecurve_enter_notify(GtkWidget *widget, GdkEventCrossing *event, gpoint
 {
   dt_iop_module_t *self = (dt_iop_module_t *)user_data;
   dt_iop_basecurve_gui_data_t *c = (dt_iop_basecurve_gui_data_t *)self->gui_data;
-  c->mouse_x = fabsf(c->mouse_x);
-  c->mouse_y = fabsf(c->mouse_y);
+  c->mouse_x = fabs(c->mouse_x);
+  c->mouse_y = fabs(c->mouse_y);
   gtk_widget_queue_draw(widget);
   return TRUE;
 }
@@ -427,8 +427,8 @@ dt_iop_basecurve_leave_notify(GtkWidget *widget, GdkEventCrossing *event, gpoint
   dt_iop_module_t *self = (dt_iop_module_t *)user_data;
   dt_iop_basecurve_gui_data_t *c = (dt_iop_basecurve_gui_data_t *)self->gui_data;
   // sign swapping for fluxbox
-  c->mouse_x = -fabsf(c->mouse_x);
-  c->mouse_y = -fabsf(c->mouse_y);
+  c->mouse_x = -fabs(c->mouse_x);
+  c->mouse_y = -fabs(c->mouse_y);
   gtk_widget_queue_draw(widget);
   return TRUE;
 }
